+++ thebleep/rules/dirty_untar.py
@@ -32,7 +32,8 @@ def match(command):
 
 def get_new_command(command):
     dir = shell.quote(_tar_file(command.script_parts)[1])
-    return shell.and_('mkdir -p {dir}', '{cmd} -C {dir}') \
+    return shell.and_('{} {{dir}}'.format(shell.mkdir_command()),
+                      '{cmd} -C {dir}') \
         .format(dir=dir, cmd=command.script)